Add files via upload
poprawki i debugowanie kolizji

+++ src/Entity/Enemy.java
@@ -2,7 +2,7 @@ package Entity;
 
 public class Enemy extends Mob{
     public Enemy(int hp, float radius, float angle, float[] position,
-                 int movementSpeed, boolean canShoot, int attackFrequency, int damage) {
+                 float movementSpeed, boolean canShoot, int attackFrequency, int damage) {
         super(hp, radius, angle, position, movementSpeed, canShoot, attackFrequency, damage);
     }
     public Enemy() {
